backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
python-multipart
python-jose[cryptography]
requests
pydantic
email-validator
google-auth
google-auth-oauthlib
google-api-python-client
python-docx
pdfplumber
alembic
python-dateutil
